refactor: group event definitions by model type (#14790)
